requirements.txt
# Pin versions as the project stabilizes.
numpy>=1.24
pandas>=2.0
pyarrow>=14.0
netcdf4>=1.6
xarray>=2023.1
pyyaml>=6.0
python-dotenv>=1.0
jupyter>=1.0
pytest>=7.0

# Optional for app deployment (uncomment as needed)
# fastapi>=0.100
# uvicorn[standard]>=0.22
# streamlit>=1.28
